DL-137 removed redundant code
