--- recipes/mapr_cldb.rb
@@ -1,3 +1,0 @@
-log "\n=========== Start MapR mapr_cldb.rb =============\n"
-
-package 'mapr-cldb'
