--- recipes/mapr_historyserver.rb
@@ -1,3 +1,0 @@
-log "\n=========== Start MapR mapr_historyserver.rb =============\n"
-
-package 'mapr-historyserver'
