--- recipes/mapr_nodemanager.rb
@@ -1,3 +1,0 @@
-log "\n=========== Start MapR mapr_nodemanager.rb =============\n"
-
-package 'mapr-nodemanager'
