--- recipes/mapr_resourcemanager.rb
@@ -1,3 +1,0 @@
-log "\n=========== Start MapR mapr_resourcemanager.rb =============\n"
-
-package 'mapr-resourcemanager'
