--- recipes/mapr_zookeeper.rb
@@ -1,3 +1,0 @@
-log "\n=========== Start MapR mapr_zookeeper.rb =============\n"
-
-package 'mapr-zookeeper'
